+++ test/unit/spec/util/StringSpec.js
@@ -142,5 +142,38 @@ define(['util/string'], (StringUtil) => {
                 expect(StringUtil.isEmptyString(testCase)).toBeFalse();
             });
         });
+
+        [
+            ['abcd', 'abcd'],
+            ['ab_cd', 'ab_cd'],
+            ['ab cd', 'ab_cd'],
+            ['123', 'obj_123'],
+            ['ab&!*()', 'ab_____'],
+            ['a/b/c/abcd', 'a_b_c_abcd'],
+            ['123_a.b|c-d_EFG', '123_a.b|c-d_EFG'],
+            ['_|.---', '_|.---'],
+            [
+                'ThisIsAnOtherwiseValidButVeryLongNameThatShouldVeryMuchBeTruncatedToBeLessThan255CharactersButWhichMeansIMustTypeAndTypeAndTypeAndTypeForeverAndEverUtilIGetToTheRightLengthICouldImplementASimpleStringGeneratorThatPutsIn1000AsOrSomethingButThisIsOddlySoothingAndWillMakeAnAmusingPRIHope',
+                'ThisIsAnOtherwiseValidButVeryLongNameThatShouldVeryMuchBeTruncatedToBeLessThan255CharactersButWhichMeansIMustTypeAndTypeAndTypeAndTypeForeverAndEverUtilIGetToTheRightLengthICouldImplementASimpleStringGeneratorThatPutsIn1000AsOrSomethingButThisIsOddlySooth',
+            ],
+        ].forEach((testCase) => {
+            it('should sanitize a name for workspace consumption', () => {
+                const [input, expected] = testCase;
+                expect(StringUtil.sanitizeWorkspaceObjectName(input)).toEqual(expected);
+            });
+        });
+
+        [
+            ['a/b/c/abcd', 'abcd'],
+            ['/a/b/ab_cd', 'ab_cd'],
+            ['a/b/c/d/e/f/g/ab cd', 'ab_cd'],
+            ['a/b/c/123', 'obj_123'],
+            ['a/b/c/1_a.B|C-d', '1_a.B|C-d'],
+        ].forEach((testCase) => {
+            it('should sanitize a file path for workspace consumption', () => {
+                const [input, expected] = testCase;
+                expect(StringUtil.sanitizeWorkspaceObjectName(input, true)).toEqual(expected);
+            });
+        });
     });
 });